Extract the bits from my Plack patch.
miyagawa thinks this would be better off in its own dist, so let's try
that. It'll be a little irritating having to duplicate the ::Negotiate
logic, but it's not like I haven't written that once myself already. One
advantage is I get to steal Apache's names...

For now, keep all this in one distribution. It'll end up with a lot of
deps, but don't worry about that until it's actually a problem for
someone.