Redo the AuthZ design.
There are now two main classes: AuthZ itself, which represents an ACL,
and AuthZ::Base and subclasses, which represent entries in that list.

ACEs have an ->on_match hook, and if their condition is true they send
control that way instead of to ->app. ACLs have ->on_allow and ->on_deny
hooks, defaulting to ->app and a 401 respectively, and a list of ACEs.
At ->prepare_app time the ACL runs through the list and wires up all the
ACEs, to each other and to the ACL's hooks. This means at run time it's
got nothing to do except call the first ACE.

This commit adds the following ACE types:

    AuthZ::All      always matches
    AuthZ::User     requires $env->{REMOTE_USER} to be defined
                    (renamed from ::RequireUser)

User should eventually learn to match the username against various
things (regexp, arrayref, hashref, subref).

This commit also adds a Util::Static middleware, which I'm not entirely
sure about: for one thing, it isn't really a middleware at all. However,
it *is* necessary, under some name, and the ::Middleware prefix seems to
be obligatory to make all the magic work.